Fix the underline highlight on paper-input element
The rightmost part of the underline on a paper-input remained
highlighted after the paper-input lost focus. This was broken due to my
recent changes to paper-styles, I fixed this in my fork of paper-input,
so we'll have to use the fork until the Polymer team gets rid of
shadow DOM piercing CSS rules.