Add the ability to create a new debug configuration from debug-toolbar
The first entry in the configurations dropdown in the debug-toolbar is
now "New...", selecting this entry will pop up a dialog that lets the
user enter the name for a new debug configuration and select the debug
engine is will apply to. After those details are provided the new
engine-specific debug configuration will be displayed for editing.