Add index.d.ts to polymer-elements typings
The index file includes all the individual element declarations to make
it easier to consume the typings in packages that depend on this one.
To streamline this use case the index file is also referenced in
package.json so that running `tsd link` in a package that depends on
this one will automatically add a reference to polymer-elements index
file to the .d.ts file managed by tsd.