Generate proper external module declarations instead of ambient ones
dtsGenerator is no longer used to generate the type definitions for this
package, it will be removed from the dependencies list in the near
future if the "proper external module" format doesn't result in any
unforeseen issues.

This change also made it necessary to re-introduce the `src` directory
for `.ts` source files due to an issue with the TypeScript compiler
<https://github.com/Microsoft/TypeScript/issues/4667> where TypeScript
would attempt to build the `.ts` files in this package while building
a project that consumes this package (instead of simply referencing the
`.d.ts` files in this package). The workaround for now is to ensure the
`.d.ts` files `.ts` files are in separate directories.